+++ PKHeXMac/Core/Items/ItemStorage2.swift
@@ -13,7 +13,7 @@ struct ItemStorage2: ItemStorage {
     static let instanceGS = ItemStorage2(version: GameVersion.GS)
     static let instanceC = ItemStorage2(version: GameVersion.C)
 
-    private static let pouchItems: [Int] = [
+    private static let pouchItems: [UInt16] = [
         003, 008, 009, 010, 011, 012, 013, 014, 015, 016, 017, 018, 019,
         020, 021, 022, 023, 024, 026, 027, 028, 029, 030, 031, 032, 033,
         034, 035, 036, 037, 038, 039, 040, 041, 042, 043, 044, 046, 047,
@@ -26,18 +26,18 @@ struct ItemStorage2: ItemStorage {
         170, 172, 173, 174, 180, 181, 182, 183, 184, 185, 186, 187, 188, 189
     ]
 
-    private static let pouchBall: [Int] = [
+    private static let pouchBall: [UInt16] = [
         1, 2, 4, 5, 157, 159, 160, 161, 164, 165, 166
     ]
 
-    private static let pouchKeyGS: [Int] = [
+    private static let pouchKeyGS: [UInt16] = [
         007, 054, 055, 058, 059, 061, 066, 067, 068, 069, 071, 127, 128,
         130, 133, 134, 175, 178
     ]
 
-    private static let pouchKeyC: [Int] = pouchKeyGS + [070, 115, 116, 129]
+    private static let pouchKeyC: [UInt16] = pouchKeyGS + [070, 115, 116, 129]
 
-    private static let pouchTMHM: [Int] = [
+    private static let pouchTMHM: [UInt16] = [
         191, 192, 193, 194, 196, 197, 198, 199, 200, 201, 202, 203, 204,
         205, 206, 207, 208, 209, 210, 211, 212, 213, 214, 215, 216, 217,
         218, 219, 221, 222, 223, 224, 225, 226, 227, 228, 229, 230, 231,
@@ -54,7 +54,7 @@ struct ItemStorage2: ItemStorage {
         return true
     }
 
-    func getItems(type: InventoryType) -> [Int] {
+    func getItems(type: InventoryType) -> [UInt16] {
         switch type {
         case .Items:
             return ItemStorage2.pouchItems
